requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gelfHandler==1.0.0
graypy==2.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
waitress==2.0.0
Werkzeug==2.0.1
